backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.26
python-dotenv>=1.0.0
httpx>=0.26.0
numpy>=1.26.0
scikit-learn>=1.4.0
pydantic>=2.5.0
